+++ components/modal/index.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState, useCallback } from 'react'
-import { unmountComponentAtNode, createPortal } from 'react-dom'
+import { unmountComponentAtNode, createPortal, render } from 'react-dom'
 import { CSSTransition } from 'react-transition-group'
 import Classnames from 'classnames'
 import Provider from '../context/index'
@@ -114,90 +114,143 @@ const InternalModalComp = ({
         className={Classnames(`${PREFIX}__mask`, {
           [`${PREFIX}__mask--visible`]: visible
         })}
+      />
+      <div
+        className={`${PREFIX}__outter`}
+        style={{ display: vi === false && 'none' }}
         onClick={() => {
           if (maskClosable && onCancel) {
             onCancel()
           }
         }}
-      />
-      <div
-        className={`${PREFIX}__container`}
-        style={{ display: vi === false && 'none' }}
-        tabIndex={-1}
-        ref={modalRef}
-        onKeyDown={trapTabKey}
       >
-        <CSSTransition
-          in={visible}
-          timeout={0}
-          classNames={'modal-transition'}
-          onExited={() => {
-            setTimeout(() => {
-              setVi(false)
-              focusedElementBeforeOpenModal.current.focus()
-            }, 300)
-          }}
-        >
-          <div
-            className={Classnames(`${PREFIX}__wrapper`, `${PREFIX}__wrapper--${size}`)}
-            style={{ width, height, ...style }}
+        <div className={`${PREFIX}__container`} tabIndex={-1} ref={modalRef} onKeyDown={trapTabKey}>
+          <CSSTransition
+            in={visible}
+            timeout={0}
+            classNames={'modal-transition'}
+            onExited={() => {
+              setTimeout(() => {
+                setVi(false)
+                focusedElementBeforeOpenModal.current.focus()
+              }, 300)
+            }}
           >
             <div
-              className={Classnames(`${PREFIX}__header`, {
-                [`${PREFIX}__header--divided`]: showHeaderDivider
-              })}
+              className={Classnames(`${PREFIX}__wrapper`, `${PREFIX}__wrapper--${size}`)}
+              style={{ width, height, ...style }}
+              onClick={(e) => {
+                e.stopPropagation()
+              }}
             >
-              {title}
-              <Icon
-                name={'close'}
-                style={{ cursor: 'pointer' }}
-                onClick={() => {
-                  if (onCancel) {
-                    onCancel()
-                  }
-                }}
-              />
-            </div>
-            <div className={`${PREFIX}__content`}>{children}</div>
-            {footer !== null && (
               <div
-                className={Classnames(`${PREFIX}__footer`, {
-                  [`${PREFIX}__footer--divided`]: showFooterDivider
+                className={Classnames(`${PREFIX}__header`, {
+                  [`${PREFIX}__header--divided`]: showHeaderDivider
                 })}
               >
-                {footer === undefined && cancelText !== null && (
-                  <Button
-                    type={'line'}
-                    onClick={() => {
-                      if (onCancel) {
-                        onCancel()
-                      }
-                    }}
-                  >
-                    {cancelText || localeDatas.modal.cancelText}
-                  </Button>
-                )}
-                {footer === undefined && confirmText !== null && (
-                  <Button
-                    type={'primary'}
-                    onClick={() => {
-                      if (onConfirm) {
-                        onConfirm()
-                      }
-                    }}
-                  >
-                    {confirmText || localeDatas.modal.confirmText}
-                  </Button>
-                )}
-                {footer}
+                {title}
+                <Icon
+                  name={'close'}
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => {
+                    if (onCancel) {
+                      onCancel()
+                    }
+                  }}
+                />
               </div>
-            )}
-          </div>
-        </CSSTransition>
+              <div className={`${PREFIX}__content`}>{children}</div>
+              {footer !== null && (
+                <div
+                  className={Classnames(`${PREFIX}__footer`, {
+                    [`${PREFIX}__footer--divided`]: showFooterDivider
+                  })}
+                >
+                  {footer === undefined && cancelText !== null && (
+                    <Button
+                      type={'line'}
+                      onClick={() => {
+                        if (onCancel) {
+                          onCancel()
+                        }
+                      }}
+                    >
+                      {cancelText || localeDatas.modal.cancelText}
+                    </Button>
+                  )}
+                  {footer === undefined && confirmText !== null && (
+                    <Button
+                      type={'primary'}
+                      onClick={() => {
+                        if (onConfirm) {
+                          onConfirm()
+                        }
+                      }}
+                    >
+                      {confirmText || localeDatas.modal.confirmText}
+                    </Button>
+                  )}
+                  {footer}
+                </div>
+              )}
+            </div>
+          </CSSTransition>
+        </div>
       </div>
     </div>,
     container || defaultContainer.current
   )
+}
+
+const confirmIconMap = {
+  success: { name: 'check-circle', color: '#1DA653' },
+  error: { name: 'close-circle', color: '#EB5252' },
+  warning: { name: 'exclamation-circle', color: '#e19d0b' },
+  info: { name: 'info-circle', color: '#4284F5' }
+}
+
+const confirm = ({ onConfirm, onCancel, title = '提示', content, type = 'default', confirmText, cancelText }) => {
+  const confirmContainer = document.createElement('div')
+
+  document.body.appendChild(confirmContainer)
+  const modal = React.createElement(ModalComp, {
+    container: confirmContainer,
+    title,
+    width: 480,
+    height: 240,
+    visible: true,
+    confirmText,
+    cancelText,
+    onConfirm: () => {
+      onConfirm && onConfirm()
+      confirmContainer.parentNode.style.removeProperty('overflow')
+      unmountComponentAtNode(confirmContainer)
+      confirmContainer.parentNode.removeChild(confirmContainer)
+    },
+    showFooterDivider: false,
+    children: (
+      <div style={{ display: 'flex', flex: 1 }}>
+        {type !== 'default' && (
+          <Icon
+            name={confirmIconMap[type] && confirmIconMap[type].name}
+            style={{
+              color: confirmIconMap[type] && confirmIconMap[type].color,
+              fontSize: '48px',
+              marginRight: 12
+            }}
+          />
+        )}
+        {content}
+      </div>
+    ),
+    onCancel: () => {
+      onCancel && onCancel()
+      confirmContainer.parentNode.style.removeProperty('overflow')
+      unmountComponentAtNode(confirmContainer)
+      confirmContainer.parentNode.removeChild(confirmContainer)
+    }
+  })
+  render(modal, confirmContainer)
 }
 
 const ModalComp = Provider(InternalModalComp)