+++ components/select/MultipleInput.js
@@ -15,7 +15,6 @@ const MultipleInput = ({
   multipleMode = 'nowrap',
   onFocus,
   theme,
-  onBlur,
   onClick,
   onClickOption,
   onClear,
@@ -111,7 +110,7 @@ const MultipleInput = ({
         )}
         {searchable && !disabled && (
           <div className="hi-select__input--search">
-            <input type="text" onFocus={onFocus} onBlur={onBlur} readOnly className="hi-select__input-readOnly" />
+            <input type="text" onFocus={onFocus} readOnly className="hi-select__input-readOnly" />
           </div>
         )}
       </div>