+++ docs/demo/select/section-type.jsx
@@ -171,6 +171,44 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['受控']
+  },
+  {
+    code: `import React from 'react'
+import Select from '@hi-ui/hiui/es/select'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      value: '3',
+      singleList: [
+        { title:'电视', id:'3', disabled: true },
+        { title:'手机', id:'2' },
+        { title:'笔记本', id:'4', disabled: true },
+        { title:'生活周边', id:'5' },
+        { title:'办公', id:'6' },
+      ]
+    }
+  }
+
+  render () {
+    const { value, singleList } = this.state
+    return (
+      <Select
+        type='single'
+        legacyV2
+        style={{ width: 200 }}
+        data={singleList}
+        value={value}
+        onChange={ids => {
+          this.setState({
+            value: ids
+          })
+        }}
+      />
+    )
+  }
+}`,
+    opt: ['v2']
   }
 ]
 const DemoType = () => (